chore: Reorganize dockerfile and deps
Attempt to optimize time consumed building containers in CI
