+++ .dockerignore
@@ -6,3 +6,5 @@ doc
 tools/
 skaffold.yaml
 target/
+*.md
+*.yaml
