+++ .github/workflows/ci.yml
@@ -22,6 +22,7 @@ env:
   NEXTEST_RETRIES: 3
   KEYSTONE_DEV_KEK: 4242424242424242424242424242424242424242424242424242424242424242
   KEYSTONE_ALLOW_ENV_KEK: "1"
+  SPIRE_VERSION: "1.15.1"
 
 jobs:
   test:
@@ -56,6 +57,14 @@ jobs:
         with:
           tool: cargo-nextest
 
+      # - name: Install SPIRE server
+      #   run: |
+      #     wget https://github.com/spiffe/spire/releases/download/v${SPIRE_VERSION}/spire-${SPIRE_VERSION}-linux-amd64-musl.tar.gz
+      #     tar -xvf spire-${SPIRE_VERSION}-linux-amd64-musl.tar.gz
+      #     # Move binaries to a standard directory and add to PATH
+      #     sudo cp spire-${SPIRE_VERSION}/bin/spire-server /usr/local/bin/
+      #     sudo cp spire-${SPIRE_VERSION}/bin/spire-agent /usr/local/bin/
+
       - name: Rust Cache
         uses: swatinem/rust-cache@98c8021b550208e191a6a3145459bfc9fb29c4c0 # v2.8.0
 
@@ -73,6 +82,9 @@ jobs:
 
       - name: Run tests for raft drivers
         run: cargo nextest run -p test_integration --profile raft --no-capture
+
+      # - name: Run Rest API v3 tests
+      #   run: cargo nextest run -p test_api --profile api --no-capture --test integration_api_v3
 
       - name: Run Doc tests
         run: cargo test --doc
