+++ tests/api/Cargo.toml
@@ -14,6 +14,7 @@ name = "test_api"
 path = "src/lib.rs"
 
 [dependencies]
+async-trait.workspace = true
 base64.workspace = true
 bytes.workspace = true
 eyre.workspace = true
@@ -21,7 +22,7 @@ derive_builder.workspace = true
 http.workspace = true
 openstack-keystone-api-types = { workspace = true, features = ["builder"] }
 openstack-keystone-distributed-storage = { workspace = true }
-openstack_sdk = { version = "0.22" }
+openstack_sdk = { version = "0.22", default-features = false, features = ["async", "identity" ] }
 reqwest = { workspace = true, features = ["json", "http2", "multipart", "gzip", "deflate"] }
 secrecy = { workspace = true, features = ["serde"] }
 serde.workspace = true
@@ -39,7 +40,6 @@ webauthn-authenticator-rs = { workspace = true, features = ["softtoken"] }
 x509-parser = "0.16"
 url = { workspace = true }
 uuid = { workspace = true, features = ["v4"] }
-async-trait.workspace = true
 
 [[test]]
 name = "integration_api_v3"
