+++ tests/api/tests/api_v3/auth/project.rs
@@ -15,15 +15,49 @@
 use std::sync::Arc;
 
 use eyre::Result;
+use tracing_test::traced_test;
 
+use openstack_keystone_api_types::v3::user::UserCreateBuilder;
 use openstack_sdk::{AsyncOpenStack, config::CloudConfig};
 
 use test_api::auth::project::*;
+use test_api::common::get_session_by_user_password;
+use test_api::guard::ResourceGuard;
+use test_api::identity::user::create_user;
+use uuid::Uuid;
 
 #[tokio::test]
 async fn test_list_user_projects() -> Result<()> {
     let test_client = Arc::new(AsyncOpenStack::new(&CloudConfig::from_env()?).await?);
     let projects = list_auth_projects(&test_client).await?;
     assert!(!projects.is_empty());
+    Ok(())
+}
+
+#[tokio::test]
+#[traced_test]
+async fn test_auth_projects_empty_for_user_without_roles() -> Result<()> {
+    let tc = Arc::new(AsyncOpenStack::new(&CloudConfig::from_env()?).await?);
+    let name = format!("usr_{}", Uuid::new_v4().simple());
+    let password = "TestPassword123!";
+
+    let guard = create_user(
+        &tc,
+        UserCreateBuilder::default()
+            .name(&name)
+            .domain_id("default")
+            .enabled(true)
+            .password(password)
+            .build()?,
+    )
+    .await?;
+
+    let user_client =
+        get_session_by_user_password(&guard.name, &guard.domain_id, &password).await?;
+
+    let projects = list_auth_projects(&user_client).await?;
+    assert!(projects.is_empty());
+
+    guard.delete().await?;
     Ok(())
 }
